fix package names. Moved from `alephium.io` to their actual package names.
