+++ cache/src/main/scala/org/alephium/cache/ReadableTrie.scala
@@ -14,7 +14,9 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with the library. If not, see <http://www.gnu.org/licenses/>.
 
-package org.alephium.io
+package org.alephium.cache
+
+import org.alephium.io.IOResult
 
 trait ReadableTrie[K, V] {
   def get(key: K): IOResult[V]
