+++ cache/src/main/scala/org/alephium/cache/SparseMerkleTrie.scala
@@ -14,12 +14,14 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with the library. If not, see <http://www.gnu.org/licenses/>.
 
-package org.alephium.io
+package org.alephium.cache
 
 import akka.util.ByteString
 
 import org.alephium.crypto.{Blake2b => Hash}
+import org.alephium.io.{IOError, IOResult}
 import org.alephium.serde._
+import org.alephium.storage.KeyValueStorage
 import org.alephium.util.{AVector, Bytes}
 
 object SparseMerkleTrie {
