+++ cache/src/main/scala/org/alephium/cache/StagingSMT.scala
@@ -14,9 +14,11 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with the library. If not, see <http://www.gnu.org/licenses/>.
 
-package org.alephium.io
+package org.alephium.cache
 
 import scala.collection.mutable
+
+import org.alephium.io.IOResult
 
 final class StagingSMT[K, V](
     val underlying: CachedSMT[K, V],
