+++ cache/src/test/scala/org/alephium/cache/SparseMerkleTrieSpec.scala
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with the library. If not, see <http://www.gnu.org/licenses/>.
 
-package org.alephium.io
+package org.alephium.cache
 
 import scala.annotation.tailrec
 
@@ -23,7 +23,9 @@ import org.scalacheck.{Arbitrary, Gen}
 import org.scalatest.Assertion
 
 import org.alephium.crypto.{Blake2b => Hash}
+import org.alephium.io.IOError
 import org.alephium.serde._
+import org.alephium.storage.StorageFixture
 import org.alephium.util.{AlephiumSpec, AVector}
 
 class SparseMerkleTrieSpec extends AlephiumSpec {
