+++ flow/src/main/scala/org/alephium/flow/io/HeightIndexStorage.scala
@@ -19,12 +19,11 @@ package org.alephium.flow.io
 import akka.util.ByteString
 import org.rocksdb.{ReadOptions, WriteOptions}
 
-import org.alephium.flow.io.HeightIndexStorage.hashesSerde
-import org.alephium.io.{RocksDBKeyValueStorage, RocksDBSource}
-import org.alephium.io.RocksDBSource.ColumnFamily
 import org.alephium.protocol.BlockHash
 import org.alephium.protocol.model.ChainIndex
 import org.alephium.serde._
+import org.alephium.storage.rocksdb.{RocksDBKeyValueStorage, RocksDBSource}
+import org.alephium.storage.rocksdb.RocksDBSource.ColumnFamily
 import org.alephium.util.{AVector, Bytes}
 
 object HeightIndexStorage {
