+++ flow/src/main/scala/org/alephium/flow/io/Storages.scala
@@ -20,12 +20,14 @@ import java.nio.file.Path
 
 import org.rocksdb.WriteOptions
 
-import org.alephium.io.{IOResult, KeyValueSource, RocksDBKeyValueStorage, RocksDBSource}
-import org.alephium.io.RocksDBSource.ColumnFamily
-import org.alephium.io.SparseMerkleTrie.Node
+import org.alephium.cache.SparseMerkleTrie.Node
+import org.alephium.io.IOResult
 import org.alephium.protocol.Hash
 import org.alephium.protocol.config.GroupConfig
 import org.alephium.protocol.vm.WorldState
+import org.alephium.storage.KeyValueSource
+import org.alephium.storage.rocksdb.{RocksDBKeyValueStorage, RocksDBSource}
+import org.alephium.storage.rocksdb.RocksDBSource.ColumnFamily
 import org.alephium.util.AVector
 
 object Storages {
