+++ flow/src/main/scala/org/alephium/flow/io/TxStorage.scala
@@ -21,8 +21,14 @@ import org.rocksdb.{ReadOptions, WriteOptions}
 import org.alephium.flow.core.BlockChain
 import org.alephium.flow.core.BlockChain.{TxIndex, TxIndexes}
 import org.alephium.io._
-import org.alephium.io.RocksDBSource.ColumnFamily
 import org.alephium.protocol.Hash
+import org.alephium.storage.KeyValueStorage
+import org.alephium.storage.rocksdb.{
+  RocksDBKeyValueCompanion,
+  RocksDBKeyValueStorage,
+  RocksDBSource
+}
+import org.alephium.storage.rocksdb.RocksDBSource.ColumnFamily
 import org.alephium.util.AVector
 
 trait TxStorage extends KeyValueStorage[Hash, TxIndexes] {
