+++ protocol/src/test/scala/org/alephium/protocol/vm/VMFactory.scala
@@ -16,8 +16,9 @@
 
 package org.alephium.protocol.vm
 
-import org.alephium.io.{SparseMerkleTrie, StorageFixture}
+import org.alephium.cache.SparseMerkleTrie
 import org.alephium.protocol.Hash
+import org.alephium.storage.StorageFixture
 
 trait VMFactory extends StorageFixture {
   lazy val cachedWorldState: WorldState.Cached = {
