+++ storage/src/main/scala/org/alephium/storage/KeyValueSource.scala
@@ -14,7 +14,9 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with the library. If not, see <http://www.gnu.org/licenses/>.
 
-package org.alephium.io
+package org.alephium.storage
+
+import org.alephium.io.IOResult
 
 trait KeyValueSource {
   def close(): IOResult[Unit]
