+++ storage/src/main/scala/org/alephium/storage/KeyValueStorage.scala
@@ -14,11 +14,13 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with the library. If not, see <http://www.gnu.org/licenses/>.
 
-package org.alephium.io
+package org.alephium.storage
 
 import akka.util.ByteString
 
-import org.alephium.serde._
+import org.alephium.io.IOResult
+import org.alephium.io.IOUtils
+import org.alephium.serde.*
 
 trait AbstractKeyValueStorage[K, V] {
