+++ storage/src/main/scala/org/alephium/storage/RawKeyValueStorage.scala
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with the library. If not, see <http://www.gnu.org/licenses/>.
 
-package org.alephium.io
+package org.alephium.storage
 
 import akka.util.ByteString
